svg/animations/repeat-iteration-event-005.svg

<svg xmlns="http://www.w3.org/2000/svg"><title>No repeat iteration at interval end</title><script xmlns="http://www.w3.org/1999/xhtml" src="/resources/testharness.js"/><script xmlns="http://www.w3.org/1999/xhtml" src="/resources/testharnessreport.js"/><rect width="100" height="100" fill="green"><animate id="anim" attributeName="visibility" begin="0s" dur="2s" repeatCount="4" to="visible"/><set attributeName="width" begin="anim.repeat(4)" to="0"/></rect><script>function waitForFrame() { return new Promise(resolve =&gt; { window.requestAnimationFrame(resolve); }); } promise_test(t =&gt; { document.documentElement.setCurrentTime(7.995); let set = document.querySelector(&quot;set&quot;); new EventWatcher(t, set, &apos;beginEvent&apos;); let anim = document.getElementById(&quot;anim&quot;); let animWatcher = new EventWatcher(t, anim, &apos;endEvent&apos;); return animWatcher.wait_for(&apos;endEvent&apos;).then(() =&gt; { return waitForFrame(); }).then(() =&gt; { assert_equals(set.parentNode.getBBox().width, 100); }); });</script></svg>
